Set default error callback for runTests (#47)
* Fixed - safe exit after successful run

Fixes a bug where jspmjasmine would not exit cleanly after a successful run. We now ensure that safeExit() with jasmine exit strategy is passed in the errCallback() of runTests().

* errCallback() from cli.js is set as default in runTests()

`errBack` previously in CLI is now set as the default for `runTests`. Thus users will be able to take advantage of safe exit logic when using the JS API.

* Added Support for Custom Reporters

Custom reporter can be configured in a  manner identical to Jasmine and passed using a `reporter` option in `runTests()`