BaseURL correction (#49)
* Fixed - safe exit after successful run

Fixes a bug where jspmjasmine would not exit cleanly after a successful run. We now ensure that safeExit() with jasmine exit strategy is passed in the errCallback() of runTests().

* errCallback() from cli.js is set as default in runTests()

`errBack` previously in CLI is now set as the default for `runTests`. Thus users will be able to take advantage of safe exit logic when using the JS API.

* Added Support for Custom Reporters

Custom reporters can be configured in a manner identical to Jasmine and passed using a `reporters` option in `runTests()`

* Fixed Readme Typos on Custom Reporter

* Fixes Readme Typos on Custom Reporter 2

Fixes the multiple reporters example.

* Path correction for BaseURL

If the user has specified a BaseURL in their JSPM configuration, jspm-jasmine will correct for it before importing files.